backend/pyproject.toml
[tool.poetry]
package-mode = false
name = "backend"
version = "0.1.0"
description = ""
authors = ["leocodeio <basicleo@787>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
speechrecognition = "^3.10.4"
python-dotenv = "^1.0.1"
google-generativeai = "^0.8.3"
textblob = "^0.18.0.post0"
transformers = "^4.45.2"
torch = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
